Moved ContractManager SearchTools out of Legacy
