+++ services/client/src/Apps/Contracts/ContractManager/Browser/SearchTools.tsx
